webapi/webapi-webspeech-w3c-tests/webspeech/blink/speechrecognition-start-exception.html

<!--
Test convert from Blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/speech/scripted/start-exception.html
-->

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Speech API Test: getting an exception when calling start() in the Speech JavaScript API</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var t = async_test("Test getting an exception when calling start() in the Speech JavaScript API", { timeout: 5000});

t.step(function () {
  assert_true("webkitSpeechRecognition" in self);
  assert_not_equals(webkitSpeechRecognition, null);
  window.r = new webkitSpeechRecognition();
  for (var prop in r) {
    if (prop.match('^on')) {
      r[prop] = function() {
        if (event.type === 'end') {
          r.start();
          r.abort();
          r.onend = function() {
            t.done();
          };
        } else {
          assert_throws("InvalidStateError", function () {
            r.start();
          }, "recognition has already started");
        }
      }
    }
  }
  r.start();
  assert_throws("InvalidStateError", function () {
    r.start();
  }, "recognition has already started");
});

</script>
